Organise configuration, add logger_timeout.
